apps/backend/services/auth/requirements.txt
fastapi>=0.111.0
pydantic>=2.7.0
email-validator>=2.1.1
pydantic-settings>=2.2.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
PyJWT>=2.8.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
httpx>=0.27.0
